requirements.txt
pandas
requests
#webbrowser
numpy
sqlalchemy
python-dotenv
loguru
psycopg2
holidays
pandas-market-calendars